feat: Fixes on profiles and users endpoints (#31)
* wip

* wip

* adjust conversations metadata

* wip

* remove error types

* tmp ignoring xmtpId on request typescript error

* wip

* wip
