+++ src/api/v1/authenticate.ts
@@ -82,7 +82,8 @@ authenticateRouter.post("/", async (req: Request, res: Response) => {
       .sign(new TextEncoder().encode(process.env.JWT_SECRET));
 
     res.json({ token: jwt });
-  } catch {
+  } catch (error) {
+    console.log("error:", error);
     res.status(500).json({ error: "Failed to create authentication token" });
   }
 });
