+++ src/api/v1/profiles/handlers/get-profile.ts
@@ -1,0 +1,56 @@
+import { PrismaClient } from "@prisma/client";
+import type { Request, Response } from "express";
+import type {
+  GetProfileRequestParams,
+  ProfileRequestResult,
+} from "../profiles.types";
+
+const prisma = new PrismaClient();
+
+export async function getProfile(
+  req: Request<GetProfileRequestParams>,
+  res: Response,
+) {
+  try {
+    const { xmtpId } = req.params;
+
+    if (!xmtpId) {
+      res.status(400).json({ error: "Invalid request body" });
+      return;
+    }
+
+    const profile = await prisma.profile.findFirst({
+      where: {
+        deviceIdentity: {
+          xmtpId: xmtpId,
+        },
+      },
+      include: {
+        deviceIdentity: {
+          select: {
+            xmtpId: true,
+            privyAddress: true,
+          },
+        },
+      },
+    });
+
+    if (!profile) {
+      res.status(404).json({ error: "Profile not found" });
+      return;
+    }
+
+    const profileResult: ProfileRequestResult = {
+      id: profile.id,
+      name: profile.name,
+      description: profile.description,
+      avatar: profile.avatar,
+      xmtpId: profile.deviceIdentity.xmtpId,
+      privyAddress: profile.deviceIdentity.privyAddress,
+    };
+
+    res.json(profileResult);
+  } catch {
+    res.status(500).json({ error: "Failed to fetch profile" });
+  }
+}
