+++ src/globals.d.ts
@@ -1,0 +1,5 @@
+declare namespace Express {
+  interface Locals {
+    xmtpId: string;
+  }
+}
