+++ src/middleware/noRoute.ts
@@ -1,5 +1,6 @@
 import type { Request, Response } from "express";
 
-export const noRouteMiddleware = (_req: Request, res: Response): void => {
+export const noRouteMiddleware = (req: Request, res: Response): void => {
+  console.log(`[404] No route found for ${req.method} ${req.path}`);
   res.status(404).send();
 };
